Created InitError.cpp-Moved From Main-Modularity
