+++ src/snakeProjectCpp/include/SDLC.h
@@ -1,0 +1,16 @@
+#ifndef SDLC_H
+#define SDLC_H
+
+
+class SDLC
+{
+    public:
+        SDLC();
+        virtual ~SDLC();
+
+    protected:
+
+    private:
+};
+
+#endif // SDLC_H
