+++ src/snakeProjectCpp/src/InitError.cpp
@@ -1,0 +1,26 @@
+#include <exception>
+#include <string>
+#include <SDL2/SDL.h>
+#include "InitError.h"
+
+
+InitError::InitError() :
+    exception(),
+    msg( SDL_GetError() )
+{
+}
+
+InitError::InitError( const std::string & m ) :
+    exception(),
+    msg( m )
+{
+}
+
+InitError::~InitError() throw()
+{
+}
+
+const char * InitError::what() const throw()
+{
+    return msg.c_str();
+}
