+++ src/snakeProjectCpp/src/SDLC.cpp
@@ -1,0 +1,11 @@
+#include "SDLC.h"
+
+SDLC::SDLC()
+{
+    //ctor
+}
+
+SDLC::~SDLC()
+{
+    //dtor
+}
